feat: sse EventListener 방식으로 수정, async 처리
