+++ backend/src/main/java/com/yagubogu/sse/controller/StreamController.java
@@ -5,7 +5,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 @RequiredArgsConstructor
 @RequestMapping("/api/event-stream")
@@ -15,8 +14,7 @@ public class StreamController {
 
     @GetMapping
     public ResponseEntity<Void> getEventStream() {
-        SseEmitter emitter = sseEmitterService.add();
-
-
+        sseEmitterService.add();
+        return ResponseEntity.ok().build();
     }
 }
